+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Fix mqtt code only disconnecting when connection was OK
+    - Change the schema of MQTT and SMTP credentials to make them same than those of DBMS (todo #158)
 
 ### 4.14.1